Specify postgres major version for staging
This might be good to do for the main compose file as well. Apparently
postgres version 14.0 was released very recently (2021-09-30) and for
some reason the initial database volume that was created on the staging
server was done with 13.x. This lead to the postgres container always
erroring out on startup, meaning that we could never get the bot running
properly. What a trip and a half.