Workaround for websocket errors
Apparently Slack's Bolt doesn't handle websocket disconnections
properly. This causes a non-terminating error on Node 14.x. This commit
is a workaround for that version of Node. Apparently it causes a
terminating error on Node 16.x onward, which would be okay as we've set
our container to automatically restart in our docker-compose files.