post*Message: Return promise to allow msg editing
In my testing editing messages after posting them would still cause a
notification to be sent so it's not particularly useful. This would
however allow us to keep track of automated messages and edit them as
registrations flow in throughout the day (if we ever need to).

Also cleaned up usergrouped scheduled message generation just a tiny
bit.